refactor(desktop): split visual frame document builder
- move iframe CSP, shell, theme, and bridge document construction into a dedicated visual frame document module
- keep InlineVisualFrame focused on React lifecycle, sizing, blob, and bridge event dispatch
- preserve visual frame tests and browser smoke behavior
